requirements.txt
python-dotenv
requests
openai
langchain
webdriver-manager
playwright
browser-use